Moved Nitro to the new veracruz-server way
