+++ nitro-veracruz-server/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "nitro-veracruz-server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+bincode = { version = "1.2.1", default-features = false }
+env_logger = { version = "0.9.0" }
+log = "0.4.13"
+nitro-enclave = { git = "https://github.com/veracruz-project/nitro-enclave", tag = "v0.1" }
+policy-utils = { path = "../policy-utils", features = ["std", ] }
+proxy-attestation-client = { git = "https://github.com/veracruz-project/proxy-attestation-client", tag = "v0.2" }
+structopt = { version = "0.3", features = ["wrap_help"] }
+veracruz-server = { path = "../veracruz-server", features = ["linux", ]}
+veracruz-utils = { path = "../veracruz-utils", features = ["linux", ] }
