+++ nitro-veracruz-server/src/lib.rs
@@ -1,4 +1,5 @@
-//! Veracruz Server on Specific Platforms.
+//! Lib target for Nitro Veracruz Server.
+//! It's used for unit testing
 //!
 //! ## Authors
 //!
@@ -9,5 +10,4 @@
 //! See the `LICENSE_MIT.markdown` file in the Veracruz root directory for
 //! information on licensing and copyright.
 
-#[cfg(feature = "nitro")]
-pub mod nitro;
+pub mod server;
