+++ tests/Cargo.toml
@@ -31,6 +31,7 @@ linux = [
 ]
 nitro = [
   "mbedtls/monitor_getrandom",
+  "nitro-veracruz-server",
   "policy-utils/std",
   "veracruz-server/nitro",
   "veracruz-utils/nitro",
@@ -47,9 +48,10 @@ lazy_static = "1.4.0"
 linux-veracruz-server = { path = "../linux-veracruz-server", optional = true }
 log = "0.4.13"
 mbedtls = { path = "../third-party/rust-mbedtls/mbedtls" }
-nix = "0.20.2"
-postcard = "1.0.0"
+nitro-veracruz-server = { path = "../nitro-veracruz-server", optional = true }
+nix = "0.25.0"
 policy-utils = { path = "../policy-utils", optional = true }
+postcard = "1.0.0"
 reqwest = { version = "0.11", default-features = false, features = ["blocking", "rustls-tls"]}
 transport-protocol = { path = "../transport-protocol" }
 veracruz-client = { path = "../veracruz-client" }
