+++ tests/tests/integration_test.rs
@@ -282,6 +282,8 @@ fn server_tls_loop(policy_json: String) {
     let platform_veracruz_server = linux_veracruz_server::server::VeracruzServerLinux::new(&policy_json).unwrap();
     #[cfg(feature = "icecap")]
     let platform_veracruz_server = icecap_veracruz_server::server::VeracruzServerIceCap::new(&policy_json).unwrap();
+    #[cfg(feature = "nitro")]
+    let platform_veracruz_server = nitro_veracruz_server::server::VeracruzServerNitro::new(&policy_json).unwrap();
     veracruz_server::server::server(&policy_json, platform_veracruz_server).unwrap()
 }
